requirement.txt
streamlit
langchain
langchain-google-genai
pandas
python-dotenv